+++ CHANGELOG.md
@@ -1,3 +1,28 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## [0.0.7] - 2025-02-05
+- Update Documentation
+
+## [0.0.6] - 2025-02-05
+- Update Description and Handle Ui Issue
+
+## [0.0.5] - 2025-02-02
+- Update Description and Handle Ui Issue
+
+## [0.0.4] - 2025-02-02
+- Update Documentation
+
+## [0.0.3] - 2025-02-02
+### Added
+- Initial release of the `otp_paste_field` package with customizable OTP input fields.
+- Features support for OTP length customization, paste functionality, and easy integration into Flutter apps.
+- Update And Set New Features
+
+## [0.0.2] - 2025-02-02
+### Added
+- Initial release of the `otp_paste_field` package with customizable OTP input fields.
+- Features support for OTP length customization, paste functionality, and easy integration into Flutter apps.
+
+## [0.0.1] - 2025-01-02
+### Added
+- Basic version with support for OTP input fields (no customization options).